pyproject.toml
[tool.poetry]
name = "doomcord"
version = "0.1.0"
description = "Play DOOM in Discord using ASCII art rendering"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^2.0.0"
numpy = "^1.21.0"
pillow = "^9.0.0"
python-dotenv = "^0.19.0"
aiohttp = "^3.8.0"
pygame = "^2.1.0"
python-doom = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.10.0"
flake8 = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
doomcord = "src.bot.main:main"
